dpaa_eth: Set number of tx queues to NR_CPUS
commit fd8ad2ad54cd78f31a115550c957429f2b7dbdfa from
https://git.freescale.com/git-private/cgit.cgi/ppc/t4240/linux.git

The TX code for dpaa_eth relies on having one tx queue for each
cpu in the system. The problem is that we hard-coded the number
of TX queues to be 8, which was "fine" when we only had up to 8
cores, but problematic when we have 24. NR_CPUS will ensure that
we never have fewer queues than cores.

Signed-off-by: Andy Fleming <afleming@freescale.com>