dpaa_eth: Set number of tx queues to NR_CPUS
The TX code for dpaa_eth relies on having one tx queue for each
cpu in the system. The problem is that we hard-coded the number
of TX queues to be 8, which was "fine" when we only had up to 8
cores, but problematic when we have 24. NR_CPUS will ensure that
we never have fewer queues than cores.

Signed-off-by: Andy Fleming <afleming@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>