powerpc/fsl: add the usb device a little later
The swiotlb_setup_bus_notifier is register via arch_initcall, and
the fsl usb device nodes is also added with arch_initcall. So the
usb device may be registered before the swiotlb bus notifier, and
then it will be not set to the correct dma mask. Add this device
a little later.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>