powerpc/85xx: miscellaneous P5040 device tree fixes
Most of these are because of a bad copy/paste from the P5020 device tree.

1) Remove the error interrupts from the localbus node.

2) Fix the localbus compatible property.  Now it matches upstream.

3) Add "fsl,qoriq-pcie-v2.4" as a compatible string to the PCI nodes,
because the P5040 PCIe controllers are v2.4.  We keep the v2.2 string
to provide backwards-compatibility with older kernels.

4) Remove the compatible property update for the SATA nodes, because that
is used to trigger a SATA erratum work-around that the P5040 does not need.

Signed-off-by: Timur Tabi <timur@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2.1-20120824-yocto.tar.gz tarball.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>