dpaa_eth: Optimize construction of the linear part of the skbuff
When constructing a skbuff out of an incoming contiguous frame, the
skbuff's linear part must contain at least the protocol headers (the
kernel expects them there). We used to memcopy a chunk of predefined
size, which was unnecessarily large.
When the Parse Results are available, we can limit the size of the
memcopied buffer to strictly the protocol headers lenght. This boosts
the bandwidth of a netperf TCP_SENDFILE test by about 10%, on the S/G
variant of the DPAA-Ethernet driver.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.3-SOURCE-20121114-yocto.iso tarball.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>