dpaa_eth: Reserve extra headroom in the case of SG FDs
In the non-SG codebase, we instruct FMan to reserve the appropriate
number of headroom bytes such that an Rx frame can be forwarded with
possibly some extra headers added en route.

In the SG code, we have to manage the skbuff ourselves. Adding the extra
headroom required for the described scenario.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.3-SOURCE-20121114-yocto.iso tarball.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>