kgdb: Implement kgdb over ethernet with NET_POLL
At one point this was very similar to the version Matt Mackall wrote
to allow for the use of KGDB over ethernet.  Since then it has been
reworked to fit into the unified KGDB model.

This kgdb I/O driver can work as a kernel built-in or a kernel
module.

[sonic.zhang@analog.com: Fix kconfig dependency to NET_POLL_CONTROLLER]
[dongdong.deng@windriver.com: fix error in cleanup_kgdboe()]
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>