kgdboc, tty: use tty open to start low level drivers
This patch adds two new hooks to the tty layer in tty_io.c to allow
kgdboc to open a tty prior to the start of the user space processes,
as well as to open a tty to serial port which has nothing connected on
it.  This is for the purpose of starting the low level serial drivers
such that a control-c can be caught via the RX poll mechanism.

The two new functions, tty_console_poll_open() and
tty_console_poll_close() take care of setting/cleaning up an empty
"struct file *filp" as well as keeping the reference counts correct
for the tty device.  The new function are only accessible if
CONFIG_CONSOLE_POLL as it is a direct extension to the console poll
API.  All the filp operation for the new functions take place in the
tty_io.c so no further functions have to be exported externally as
well as allowing for the internals of the tty layer to remain private.

A few minor changes were needed in tty_io.c to deal with the fact that
the inode and fpath.dentry were NULL with the generic flip structure.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>