dts/fsl: Increase range of dynamic QMan fqids
The DPAA Ethernet driver uses dynamic fqids for the default and
error Rx/Tx queues. Since the number of Tx queues was increased to
a maximum of 24 on T4 platform, not enough fqids remain available
for all ports on T4, leading to failed probing of the last few ones.

Double the size of the pool from which dynamic fqids are allocated
to solve this problem.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>