gianfar: Adding TX interrupts handling for waking up the stopped TX queues
When RX_TX_BUFF_XCHG is enabled, stopped TX queues are not enabled again
as TX interuupts are disabled and indication that tx queue is ready for transmission
is not received. Enabling TX interrupt to re-enable the TX queues if disabled.
TX interrupts are masked by default and are enabled only if TX queues are stopped.

Signed-off-by: Rajan Gupta <rajan.gupta@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>