net/dpa: initialize the sg table before using it
According to the DPAA manual, the reserved bits in sg table
must be 0. So we should initialize all the reserved bits in sg
table to 0 before using it. This also fix a bug where the simics
would interpret the buffer address as 48bit (40 bit in fact) and
then get a wrong buffer address if there are some noise in the
first reserve area.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>