dpaa_eth: Revert "dpaa_eth: Use standard printing API"
This reverts commits:
 - 4e68237dbbf6bea3e0cd926cbad2a182d75e9dff "dpaa_eth:
Use standard printing API" and
 - e3e44370d93eb93c3307d1d17a3932552bc2dd0a "dpaa_eth: Fix compilation
error with FSL_DPAA_1588".

For esoteric reasons, what was supposed to be a purely cosmetic patch
turned out to incur a 3-5% performance penalty on the DPAA-Ethernet
driver (both SG and non-SG variants), on parts with (or reduced to) 1 or
2 cores. We have no choice but reverting this for now.

The two commits referred to here should have been originally a single
commit, so squashing and reverting both in one step.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.3-SOURCE-20121114-yocto.iso tarball. Use
raw_smp_processor_id for cpu_printk macro.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>