mips64: Add support for generic MIPS64R2 processor
QEMU for 64-bit MIPS provides a single MIPS64R2 processor selection:
MIPS64R2-generic, reporting Processor ID 0x00010000. This encoding
corresponds to the MIPS Technologies (MTI) code in the company field
with the Processor ID and Revision subfields left blank.

When cpu_probe() is invoked, it dispatches to cpu_probe_mips(), which
in turn has no handler for the blank Processor ID/Revision fields, so
no valid cputype is written in the cpuinfo_mips structure, and the
boot process cannot continue.

Here we add some definitions and a switch case to enable sane handling
of this processor (pseudo)type.

Finally, adding MIPS64R2-generic support to the kernel deserves to be
recognized in the Kconfig.

Signed-off-by: Phil Staub <Phil.Staub@windriver.com>