fsl_qman: Add CEETM support
Customer Edge Egress Traffic Management (CEETM) is the new QMAN feature and
introduced from QMan v3.0. It provides the basic queuing structure - Class Queue
(CQ) which consists of a hardware managed linked list (or queue) of related
frames to replace regular FQ in egress traffic.  It supports the hierarchical
class based scheduling and traffic shaping for sub-portals on specific QMan
Direct Connect Portals (DCPs). CEETM supports up to 8 logical network interfaces
(LNI) that can each be mapped to a DCP sub-portal, and up to 32 channels that
can be configured to deliver frames to any one of the LNI. Each channel contains
a total of 16 CQs to be used in egress traffic.
This patch adds the APIs to support CEETM, allows user to allocate resources
provided by CEETM and configure the shaping and scheduling.

Signed-off-by: Haiying Wang <Haiying.Wang@freescale.com>
Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Andy Fleming <afleming@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>