msa: Don't required scheduler_ipi() hacks
From the upstream project:
git://microstate.git.sourceforge.net/gitroot/microstate/linux-msa

Rework the handling of scheduler_ipi() (and smp_call_function_interrupt()
on MIPS) to not require hacking of the parameters.  Instead, the callers
are required to handle calling the proper MSA irq_enter() and irq_exit()
functions if MSA is enabled.

Signed-off-by: Corey Minyard <cminyard@mvista.com>