powerpc: disable second thread of one core before enabling it on e6500
If Thead 1 of one core is enabled, Writing the INIA register of this
thread is ignored on e6500. Therefore, disable Thread 1 before
enabling it, especially in the case of the cpu hotplug.

In the cpu hotplug code, fsl_enable_threads() will be called to restart
Thread 1 of Core 0.

Signed-off-by: Zhao Chenhui <chenhui.zhao@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>
Signed-off-by: Andy Fleming <afleming@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>