qbman: p2041rdb/p3041ds: allocate portals, remove dts cpu-bindings.
the device trees, and changes the qbman driver initialisation model.
Portals are now "allocated" for initialisation and dynamically bound
to the required cpus. The default setup is to try to allocate a portal
for each cpu, and if that fails (there are fewer portals than cpus)
one of the cpus will share its portal with those cpus that don't have
one.

Bootargs "qportals" and "bportals" can be used to override the default
setup, however. As an example of the syntax;
       bportals=s0,1-3,s4
means that cpus 1,2,3 get their own "unshared" portals, cpus 0 and 4
get "shared" portals, and any other cpus will share from one of the
portals assigned to cpus 0 or 4, selected in a round-robin fashion.
(In this example, cpu 5 would share cpu 0's portal, cpu 6 would share
cpu4's portal, and cpu 7 would share cpu 0's portal.) If no cpus are
specified for sharing and sharing is required, one will be chosen by
the driver. If there are cpus are specified for sharing but sharing is
not required, the portals will be configured without sharing support
(this uses thinner locks).

Note that one consequence of this change is that we can't abuse the
device-tree to limit certain portals to certain pool channels for
application configuration, because we no longer know which portals will
be used for what. As such, the "fsl,qman-pool-channels" links are removed,
and the qman portal driver searches for the available pool channels
instead.

This also updates the PAMU driver to preconfigure portal stashing LIODNs
for L3 - the previous assumption of L1 assumed a cpu-assignment, which is
no longer the case.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Extracted from QorIQ-SDK-V1.3-SOURCE-20121114-yocto.iso]
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>