qbman/dpaa_eth/usdpaa: p2041rdb/p3041ds: dynamic pool-channel/CGR allocation
Pool-channels are now described via a range node+property (specifying the
range of channel IDs) rather than a separate node for each pool containing
a channel ID property. The set of valid CGR IDs are also specified via a
range node+property. Allocator APIs are added for both type of object
(in-kernel as well as for USDPAA).

Consequently, the dpaa_eth "ethernet" nodes no longer link to the pool
channel they are to use, the driver is modified to instead dynamically
allocate a single pool-channel for all its interfaces.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Extracted from QorIQ-SDK-V1.3-SOURCE-20121114-yocto.iso]
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>