powerpc/mpc85xx/e6500: work around CPU erratum A-006198
Erratum A-006198 says that a CPU can hang if a certain race
occurs between threads, involving CFX instructions (which includes
mtmsr).  The workaround is to avoid using mtmsr to disable
MSR[EE/CE/DE].  To accomplish this, we use rfmci instead,
with some gymnastics to be resistant to races that overwrite
MCSRR0/MCSRR1.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Andy Fleming <afleming@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso. Some parts don't
apply due to the context changes.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>