net/fsl_pq_mdio: add check for null tbipa
We don't need to set the tbi address for fman mdio.
So initialize the tbipa to null, and check for null
before writing the tbi address. This will drop the
hardcoded #ifndef CONFIG_FSL_FMAN and also fix the
null dereferencing oops when CONFIG_FSL_FMAN is not
enabled on DPAA SoCs.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>