Drop the scc which disable the dpaa support
The dpaa drivers is not compatible with the cpu hotplug
feature. So we introduce these scc files to disable the
dpaa support when we need enable cpu hotplug. But this method
seems not quite optimal. We should create a specific cpu hotplug
feature configuration to disable all these non-compatible drivers.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>