ocf: include <linux/irqreturn.h> in ocf-compat.h
In order to be compatible with the old kernel, the ocf-compat.h will
define the IRQ_NONE and IRQ_RETVAL to NULL if they are not defined
in other header files. So we should make sure that the file in which
these macros are defined must be included before this check. The file
"linux/irqreturn.h" was introduced only after v2.6.17, so also add
a kernel version check for this inclusion as what it did in other
part of this file.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>