FogBugz #184646: Turn on all peripheral clocks for a system reboot
commit 7aa91bb6434d72819bb71c041e1a77af4367b34a from
https://github.com/altera-opensource/linux-socfpga.git socfpga-5.15.30-lts.

When doing a software reboot, all peripheral clocks must get turned on for the
L3 interconnect to work.

This code is needed when doing a "reboot" from user-space and a peripheral
clock as been gated off. Why would a peripheral clock get gated? An example
use case would be a .ko that gets insmod and rmmod during runtime. The insmod
would turn on the IP's clock, and the rmmod would turn off the IP's clock.
Doing a "reboot" would cause the system to hang.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>

Conflicts:

	arch/arm/mach-socfpga/core.h
Signed-off-by: Liwei Song <liwei.song@windriver.com>