MLK-13638-4 regulator: fixed: add system pm routines for pinctrl
commit 7c9657fb8e7567b3d21a2abf60a3f42f37ac64af from
https://github.com/nxp-imx/linux-imx.git

At some systems, the pinctrl setting will be lost or needs to
set as "sleep" state to save power consumption. So, we need to
configure pinctrl as "sleep" state when system enters suspend,
and as "default" state after system resumes. In this way, the
pinctrl value can be recovered as "default" state after resuming.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>