drm: mipi: Bring DT format description closer to the hardware
commit 4c2a74443dc7b129b77280577d0bcb292ed62360 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Use the DSI pixel format codes as defined in the include/drm/drm_mipi_dsi.h
to describe data type for pixel format. Based on data type the incoming
pixel stream is converted to byte stream to match with the DSI requirements
detailed in sec 8.8 of the MIPI Alliance Standard for DSI specification.
This gets rid of yet another string-based format description, while
keeping the bindings simple enough and close to the hardware.

Signed-off-by: Siva Rajesh <siva.rajesh.jarugula@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Reviewed-by: Hyun Kwon <hyunk@xilinx.com
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>