MLK-14451 ARM: dts: imx7ulp: Move dts entry for sensors into another dts
commit d1f6bfca78d1e045b22df91637f545ac06b97705 from
https://github.com/nxp-imx/linux-imx

Sensors are connected to M4 and not to A-Core.
Sensors will not be exposed to A-Core via standard
i2c interface but via an i2c proxy layer over rpmsg.

Remove the dts entry to avoid the probe error messages and
add a separate dts file for the case where someone wishes to
rework the board themselves and connect sensors for testing purposes.

Signed-off-by: Adriana Reus <adriana.reus@nxp.com>
Signed-off-by: Arulpandiyan Vadivel <arulpandiyan_vadivel@mentor.com>
Signed-off-by: Srikanth Krishnakar <Srikanth_Krishnakar@mentor.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>