dt-bindings: display: fsl-imx-drm: Add i.MX8 DPR(Display Prefetch Resolve) support
commit ad20d884feada5254298a16411b0119e84cddb18 from
https://github.com/nxp-imx/linux-imx.git

The Display Prefetch Resolve(DPR) is a processor of fetching display data
before the display pipeline which needs data to drive pixels in the active
display region.  The data is transformed, or resolved from a variety of
tiled buffer formats into linear format.  The DPR transaction sequences are
issued with a high level of DRAM efficiency.  This patch adds device tree
binding doc support for i.MX8qm/qxp DPR.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>