media: xilinx: dma: Enable use of Framebuffer Write IP in V4L2 Driver
commit bf1728eea2b68a142ad56c126566851be10656cb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The Xilinx Video Framebuffer driver is video 'format aware' and
requires clients to pass video format information as part of DMA
programming. As the DMA Engine interface does not support passing
this information, a configuration object is passed through the
dma channel's private data to the Framebuffer Driver with this
information.

Signed-off-by: Jeffrey Mouroux <jeff.mouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>