MLK-16203 enable runtime pm of i2c temporary when do system suspend
commit d8da7ab89e3418e8be3f7a66ce87591b4c9e0d52 from
https://source.codeaurora.org/external/imx/linux-imx

When we do system suspend, the runtime pm will be disabled, but we need
to control the PMIC to power on/off the regulator, if the runtime pm is
disabled, if will failed to request runtime wakeup. So data transfer will
failed.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Gao Pan <pandy.gao@nxp.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>
(cherry picked from commit 93adab71408f1f162015e77715dd04ce2301d673)
[ Aisheng: fix big conflict due to
3a5ee18d2a32 ("i2c: imx: implement master_xfer_atomic callback") ]
Sign-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>