i2c: imx: implement bus recovery with gpio for Layerscape
commit f97b8a6eae07fc456fd0b870cdf2974ce94f3729 from
https://github.com/nxp-imx/linux-imx.git

Based on the I2C specification, if the data line (SDA) is stuck low,
the master should send nine clock pulses. The I2C slave device that
held the bus low should release it sometime within those nine clocks.

Because pinctrl is not supported on Layerscape, current bus recovery
is not avalible for Layerscape. This patch uses an open drain GPIO
pin to connect to the IICx_SCL to drive nine clock pulses to unlock
the I2C bus.

[Leo: This is not upstreamable unless we add pinctrl driver for the
pinmux setting and reuse the generic recovery code ]

Signed-off-by: Zhang Ying-22455 <ying.zhang22455@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Biwen Li <biwen.li@nxp.com>
[Leo: removed the -vf610- in chip specific compatible string ]
Signed-off-by: Li Yang <leoyang.li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>