MLK-17275-1 drm/bridge: adv7511: Add support for OF_DYNAMIC
commit a553d4695489478ac5c2c312a846a19b81f95cb1 from
https://github.com/nxp-imx/linux-imx

When CONFIG_OF_DYNAMIC is used, and this driver is enabled in
devicetree, but fails to probe a physical i2c client, it should disable
it's remote endpoint, so that the DRM master device won't fail to bind
the other available devices.
Usually, the remote endpoint of this device is a DRM encoder. If a DRM
encoder fails to bind, the DRM master device will also fail to bind.
This is why, we should disable the encoder node dynamically in
devicetree.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>