MLK-17590-02 driver: soc: imx: update the busfreq flow on imx8mq
commit 58f7bc0ed57e2e3df912de0cb61ccf473bf35982 from
https://github.com/nxp-imx/linux-imx

Currently, on imx8mq evk board, we only support 3200mts and 667mts
frequency setpoints. So the DDR DVFS flow need to be updated accordingly.

The dram pll and dram apb clock rate is changed in ATF when doing frequency,
in kernel side, we need to call the clk API to update the clock rate info
in clock tree.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Reviewed-by: Anson Huang <Anson.Huang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>