drm: xlnx: csc: Adding vpss-color space converter driver
commit a1849388d6491ce1179bed47f501873337c68f34 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Color space converter is part of video processing subsystem module.
It was implemented as a bridge driver connected to an encoder. The
driver implementation is based on new xlnx bridge interface. This
experimental version supports the IP default configration as RGB |
YUV444 | YUV422 | YUV420. It doesn't support the dynamic gain,
contrast and brightness.

Signed-off-by: Venkateshwar Rao G <venkateshwar.rao.gannavarapu@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>