drivers/soc/fsl: add EPU FSM configuration for deep sleep
commit 1f10deaf37c8de11318867f249839b67874e4ef2 from
https://github.com/nxp-imx/linux-imx

In the last stage of deep sleep, software will trigger a Finite
State Machine (FSM) to control the hardware procedure, such a
board isolation, killing PLLs, removing power, and so on.

When the system is waked up by an interrupt, the FSM controls
the hardware to complete the early resume procedure.

This patch configure the EPU FSM preparing for deep sleep.

Signed-off-by: Hongbo Zhang <hongbo.zhang@freescale.com>
Signed-off-by: Chenhui Zhao <chenhui.zhao@freescale.com>
Signed-off-by: Ran Wang <ran.wang_1@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>