staging: fsl_ppfe/eth: support for userspace networking
commit aaa4dcdd934f235b1893e9579d5b628f75a91626 from
https://github.com/nxp-imx/linux-imx.git

This patch adds the userspace mode support to fsl_ppfe network driver.
In the new mode, basic hardware initialization is performed in kernel, while
the datapath and HIF handling is the responsibility of the userspace.

The new command line parameter is added to initialize the ppfe module
in userspace mode. By default the module remains in kernelspace networking
mode.
To enable userspace mode, use "insmod pfe.ko us=1"

Signed-off-by: Akhil Goyal <akhil.goyal@nxp.com>
Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>