staging: fsl_ppfe/eth: handle ls1012a errata_a010897
commit  1011faa889fa8b3625e0040cafced82be66d4f01 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

On LS1012A rev 1.0, Jumbo frames are not supported as it causes
the PFE controller to hang. A reset of the entire chip is required
to resume normal operation.

To handle this errata, frames with length > 1900 are truncated for
rev 1.0 of LS1012A.

Signed-off-by: Calvin Johnson <calvin.johnson@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>