ARM: zynq: Fix spurious BUG with old DT
commit c6162c64cd6aae24823f844778bb7a025b66cafc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Commit 78604f7bad89 ("ARM: zynq: Add support for Zynq-7000S devices")
introduced a new DT node to indicate how to access efuses on Zynq 7000
devices.

Unfortunately the code did not implement the case when that node does
not exist, which is quite common in EBBR like environments where device
trees get passed down from firmware.

So let's implement the error case properly and just not access the efuse
device if no DT node for it was passed in, restoring compatibility with
older device trees.

Fixes: 78604f7bad89 ("ARM: zynq: Add support for Zynq-7000S devices")
Signed-off-by: Alexander Graf <agraf@suse.de>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>