bcm2835-dma: Add support for per-channel flags
commit  74ec72a41e1f0ae77fb74fe5826c2e355a45959e from
https://github.com/raspberrypi/linux.git rpi-5.9.y

Add the ability to interpret the high bits of the dreq specifier as
flags to be included in the DMA_CS register. The motivation for this
change is the ability to set the DISDEBUG flag for SD card transfers
to avoid corruption when using the VPU debugger.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>