MLK-19152-2 drm/imx: lcdif: force 'mode_changed' when fb width changed
commit 0f151066b220c86a62e306105a9529fd6e3897ad from
https://github.com/nxp-imx/linux-imx.git

In DRM atomic modeset check, it will not check the fb's width
change, so in later atomic commit, it will not disable the CRTC
which has no mode changed. But for LCDIF, the fb width related
registers configuration can not be done when LCDIF is running.
So force 'mode_changed' to be true when fb width changed.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(cherry picked from commit 518ff82756a39ff2d2f750596295baa4f5fca4c5)
Signed-off-by: Meng Li <Meng.Li@windriver.com>