MLK-19152-1 gpu: imx: lcdif: realize fb horizontal crop via Pigeon Mode
commit b51844af70102d84905787282c4ea3926a2203fb from
https://github.com/nxp-imx/linux-imx

According to the LCDIF specification, the Legacy Mode does not
support cropping function in the horizontal direction, so add
Pigeon Mode which can support this kind of function. And when
enable this mode, the legacy horizontal timings configuration
should use stride value but not the active width, and related
pigeon configuration should use the active width but not the
stride value.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>