MLK-19158-2 drm/imx: lcdif: improve output bus format config
commit c4cd5e8ec5834af56b3aba47f98b7d2c3aa47617 from
https://github.com/nxp-imx/linux-imx

According to LCDIF specification, the input pixel data
width and the output pixel data width can be different,
and this conversion is done by LCDIF automatically. So
config the output data width according to the requested
bus format from the encoder, instead to be same with the
input pixel data width.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>