drm/imx: sec-dsim_imx: improve DPHY TIMING configs
commit 344d0959473626faf016ac9defd19aa9dcf0ed9a from
https://github.com/nxp-imx/linux-imx

The SEC provides a table to guide the DPHY TIMINGS config based
on the PLL output bit clock frequency for DSIM. So create the
table which is used by SEC LN14LPP DPHY with HS Timing v1.2 and
this table will be used by the SEC DSIM Bridge driver to help to
config the corresponding DPHY Timings correctly for each display
mode. Along with the table, a DPHY TIMING table entry 'compare'
method is implemented for the binary search when lookup the
suitable DPHY TIMING entry.

(This is the DSIM controller driver change for DPHY TIMING config
improvements.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
[Xiaolei: The "dphy_timing_ln14lpp_v1p2" part has been submitted by
the original commit 344d09594736 "drm/imx: sec-dsim_imx: improve DPHY
TIMING configs", so this part of the code will not be modified.]
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>