drm/imx: sec-dsim_imx: improve DPHY TIMING configs
commit 40c5dc73638ef631d6a9a7c68c4eb14fc32a9c19 from
https://source.codeaurora.org/external/imx/linux-imx

The SEC provides a table to guide the DPHY TIMINGS config based
on the PLL output bit clock frequency for DSIM. So create the
table which is used by SEC LN14LPP DPHY with HS Timing v1.2 and
this table will be used by the SEC DSIM Bridge driver to help to
config the corresponding DPHY Timings correctly for each display
mode. Along with the table, a DPHY TIMING table entry 'compare'
method is implemented for the binary search when lookup the
suitable DPHY TIMING entry.

(This is the DSIM controller driver change for DPHY TIMING config
improvements.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>