serial: uartps: Change uart ID port allocation
commit f0090478ae8a76b733a846f50dc5559fff3c461f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

For IPs which have alias algorightm all the time using that alias and
minor number. It means serial10 alias ends up as ttyPS10.

If alias is not setup for probed IP instance the first unused position is
used but that needs to be checked if it is really empty because another
instance doesn't need to be probed at that time. of_alias_get_alias_list()
fills alias bitmap which exactly shows which ID is free.
If alias pointing to different not compatible IP, it is free to use.

The patch is created based on commit ae1cca3fa347 ("serial: uartps: Change
uart ID port allocation") which was later on reverted by commit
72d68197281e ("Revert "serial: uartps: Change uart ID port allocation"").
Please take a look at them for more information.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Link: https://www.mail-archive.com/linux-kernel@vger.kernel.org/msg1768840.html
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>