staging: fsl_ppfe/eth: add fixed-link support
commit df7df3175717761e42cbefefb4138c5e37b772d7 from
https://github.com/nxp-imx/linux-imx.git

In cases where MAC is not connected to a normal MDIO-managed PHY
device, and instead to a switch, it is configured as a "fixed-link".
Code to handle this scenario is added here.

phy_node in the dtb is checked to identify a fixed-link.
On identification of a fixed-link, it is registered and connected.

Signed-off-by: Calvin Johnson <calvin.johnson@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>