staging: fsl_ppfe/eth: add fixed-link support
commit fc3a91bb900644045c8c45ab3bf3e2ce82b2a181 from
https://source.codeaurora.org/external/imx/linux-imx

In cases where MAC is not connected to a normal MDIO-managed PHY
device, and instead to a switch, it is configured as a "fixed-link".
Code to handle this scenario is added here.

phy_node in the dtb is checked to identify a fixed-link.
On identification of a fixed-link, it is registered and connected.

Signed-off-by: Calvin Johnson <calvin.johnson@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>