drm: xlnx: Select VIDEOMODE_HELPERS when DRM_XLNX_PL_DISP is selected
commit 11bec21fc410d30fe2804346ac646b480418e37f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Select the VIDEOMODE_HELPERS when DRM_XLNX_PL_DISP is selected in
menuconfig. This fixes the compilation error of undefined reference to
drm_display_mode_to_videomode in the Xilinx PL Display driver.

This is also added to DRM_XLNX_MIXER as the same API is being used
there.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Saurabh Sengar <saurabh.singh@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>