soc: fsl: dpio: Fix order restoration API for QBMan 5.0
commit  19b679cfeb5cb5d63f105eab67bd4e99a2215262 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

The mechanism for indicating to HW that a frame was dropped
when performing HW order restoration changed in QBMan 5.0 to
use a management command instead of a special enqueue command.
This patch implements that change when running on a QBMan 5.0
and above device.

Signed-off-by: Roy Pledge <roy.pledge@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>