gpio: siul2-s32cc: Add GPIO driver to S32CC
commit cb4b2319ec5324f9869fc28881a45b621709f4c0 from
https://github.com/nxp-auto-linux/linux

Add GPIO driver for S32CC to allow multiple instances of the driver
to share the same register ranges due to the non-contiguous
arrangement of the EIRQ pins. The device tree binding of
this driver will be changed to use the regmap API to allow
this sharing of resources.

Issue: ALB-3170
Issue: ALB-3677
Issue: ALB-4883
Issue: ALB-5974
Signed-off-by: Matthew Nunez <matthew.nunez@nxp.com>
Signed-off-by: Phu Luu An <phu.luuan@nxp.com>
Signed-off-by: Costea Ciprian <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>