staging: xroeframer: add sysfs entries for "stats" registers
commit  cea3c585705ca7b0bd409c81ff9aeaaf8858cd0e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

This includes the first 2 registers of the stats register bank as sysfs
entries. The rest of the entries will be added on the next patch
and then the "/dev/xroefram/stats" device will be removed from
the driver's code.

Signed-off-by: Vasileios Bimpikas <vasileios.bimpikas@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>