staging: mmal-vchiq: Fix client_component for 64 bit kernel
commit  4763ae7eba06625ee1b61e6cb7f1646ad72b6b41 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

The MMAL client_component field is used with the event
mechanism to allow the client to identify the component for
which the event is generated.
The field is only 32bits in size, therefore we can't use a
pointer to the component in a 64 bit kernel.

Component handles are already held in an array per VCHI
instance, so use the array index as the client_component handle
to avoid having to create a new IDR for this purpose.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>