MLK-21150-2 drm/bridge: sec-dsim: refine hblank word count compute
commit 0bc91ab55153b58e4b96079d602f0360dae1090e from
https://source.codeaurora.org/external/imx/linux-imx

When there is no existing horizontal blanking word counts in
'dsim_hblank_par' tables, these data requires to be computed
according to the 'hfp', 'hbp' and 'hsa' timings which are in
pixel unit. So the pixel unit data requires to be converted
to word count unit data correctly to match the PLL output clk
frequency.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>