x86/resctrl: Make resctrl_mounted checks explicit
commit 8e87280d6706b5138870e28288d22901362ffb2b from
git@git.assembla.com:cavium/WindRiver.linux.git

The rdt_enable_key is switched when resctrl is mounted, and used to
prevent a second mount of the filesystem. It also enables the
architecture's RDT context switch code.

To allow the possibility of another user, this needs abstracting
to live beind a resctrl_arch_ call that says resctrl is setting
closid/rmids on tasks.

Make the resctrl_mounted checks explicit: we don't want to know if
the arch code is context switching the closid, just if resctrl is
mounted.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>