x86/resctrl: Make resctrl_arch_pseudo_lock_fn() take a plr
commit 74904865bc0e36891682aced1dad657b528e516e from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl_arch_pseudo_lock_fn() has some very arch specific behaviour,
and it takes a struct rdtgroup. This means resctrl has to expose that
structure its enums and substructures to the rest of the kernel.

The only reason resctrl_arch_pseudo_lock_fn() wants the rdtgroup is
for the closid. Embed that in the pseudo_lock_region as a hw_closid.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>