drm/imx: sec-dsim_imx: correct DSIM version register access
commit a30c02cd8b48e30c09cd33d5dad9c659365150c2 from
https://github.com/nxp-imx/linux-imx.git

During the DSIM binding stage, only enable cfg clock is not
enough to access the version register, since at this time,
if the DSIM is still in the reset state, the register read
will always return 0. So before the version register read,
the runtime pm should be in resume state.

(This is the DSIM controller driver part change.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>