drm/imx: sec-dsim_imx: correct DSIM version register access
commit 335d1436d2ce24496c7965104909c5b1886d6a55 from
https://github.com/nxp-imx/linux-imx

During the DSIM binding stage, only enable cfg clock is not
enough to access the version register, since at this time,
if the DSIM is still in the reset state, the register read
will always return 0. So before the version register read,
the runtime pm should be in resume state.

(This is the DSIM controller driver part change.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>