HSD #1806784746: EDAC/altera: Warm Reset option S10 DBE
commit a07cd1e79acf6d0f88b32b195b7ba4b60fc6860c from
https://github.com/altera-opensource/linux-socfpga.git

The Stratix10 peripheral FIFO memories can recover from double
bit errors with a warm reset instead of a cold reset.
Add the option of a warm reset for peripheral (USB, Ethernet)
memories.

CPU memories such as SDRAM and OCRAM require a cold reset for
DBEs.
Filter on whether the error is a SDRAM/OCRAM or a peripheral
FIFO memory to determine which reset to use when the warm
reset option is configured.

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>