HSD #1806784746: EDAC/altera: Warm Reset option S10 DBE
commit fed7a7691a3688d0e90f3321910bd19b7a83916c from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

The Stratix10 peripheral FIFO memories can recover from double
bit errors with a warm reset instead of a cold reset.
Add the option of a warm reset for peripheral (USB, Ethernet)
memories.

CPU memories such as SDRAM and OCRAM require a cold reset for
DBEs.
Filter on whether the error is a SDRAM/OCRAM or a peripheral
FIFO memory to determine which reset to use when the warm
reset option is configured.

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>