HSD #1806784746: EDAC/altera: Warm Reset option S10 DBE
commit 4324575356dbca0ce494a005f6a7abdf3bd29f2a from
https://github.com/altera-opensource/linux-socfpga.git socfpga-5.15.30-lts.

The Stratix10 peripheral FIFO memories can recover from double
bit errors with a warm reset instead of a cold reset.
Add the option of a warm reset for peripheral (USB, Ethernet)
memories.

CPU memories such as SDRAM and OCRAM require a cold reset for
DBEs.
Filter on whether the error is a SDRAM/OCRAM or a peripheral
FIFO memory to determine which reset to use when the warm
reset option is configured.

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>