PCI: imx: msi enable bit of rc should be set in resume
commit 7d55000d71b0150acff1442769268af9f7db4c57 from
https://source.codeaurora.org/external/imx/linux-imx

The MSI Enable bit controls delivery of MSI interrupts from components
below the Root Port.
This bit would be lost during the suspend, should be re-configured
during resume.
The PCI_MSI_FLAGS of some legacy platforms maybe mis-set to zero.
Make sure that the MSI_EN is set on them at the end of the
initialization.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>