drm/vc4: Resolve the vblank warnings on mode switching
commit  fcba3351e1de3d75d7f79e78229d42e2da99c74b from
https://github.com/raspberrypi/linux.git rpi-5.9.y

The details over when and how a driver is to service the
vblank events are sketchy, and the fkms driver was triggering
a kernel warning every time the crtc was enabled or disabled.

Copy the event handling as used by the vc4-kms driver slightly
more closely, and we avoid the warnings.

https://github.com/raspberrypi/linux/issues/3020

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>