irqchip: irq-imx-gpcv2: Disable cpuidle if no GPC_CORE_WAKE in EL3
commit 48c263bd13e046467a573305932f01b031dbe9f2 from
https://github.com/nxp-imx/linux-imx

Not all EL3 have the FSL_SIP_CONFIG_GPC_CORE_WAKE, therefore disable
the cpuidle to avoid all the cores going to sleep ending up with a
hang. This allows all the EL3 implementations to work with i.MX8MQ
even if they do not support core wake-up through GPC as a workaround.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>