clk: imx8qxp: Remove gpt0_clk to avoid warning during kernel boot up
commit d0c035a9ed5670c6aa9cfc10db17269396ff7354 from
https://source.codeaurora.org/external/imx/linux-imx

The gpt0 is assigned to ATF previously due to LPCG, context
save/restore etc. for cpu-idle feature, remove gpt0_clk to
avoid below warning during kernel boot up, if gpt0 is going
to be used in future, need to remove corresponding operations
in ATF and add it back in kernel.

[    0.291286]  gpt0: failed to power up resource 207 ret -13
[    0.291355] imx-scu-clk: probe of gpt0_clk failed with error -5

Signed-off-by: Anson Huang <Anson.Huang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>