arm64: dts: imx8qxp-mek-dsp: Change compatible string for old DSP driver
commit 60930f7370cdf06bffbb3223edb393c2b9b3c017 from
https://github.com/nxp-imx/linux-imx.git

Currently there are two Linux DSP drivers. We will use the following
compatible strings depending on which driver is used:

1) fsl,imx8qxp-dsp-v1

Used for FSL DSP Linux driver that works with the old audio framework written
completely inside NXP.

2) fsl,imx8qxp-dsp

Used for SOF Linux driver. This will be used in the future, but until
SOF Linux driver / FW will completely support all the features we will
need to support 1) and 2) in parallel.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>