drivers: base: move devices pm to tail when driver bound
commit 9ed969b2b96136e31a009c63f8bbe0af7ba3a442 from
https://github.com/nxp-imx/linux-imx.git

Only moving device pm to tail for defer probe is not enough because
the device driver may be probed out of order without defer probe
due to different driver init levels.

Let's move the device pm to tail in the driver bound to ensure
the device pm sequence is exactly the same as its probe sequence.

Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>