mfd: Add i.MX generic mix support
commit c53b4639f035b8c3467e2b63a2297bd615346666 from
https://source.codeaurora.org/external/imx/linux-imx

Some of the i.MX SoCs have a IP for interfacing the dedicated IPs with
clocks, resets and interrupts, plus some other specific control registers.
To allow the functionality to be split between drivers, this MFD driver is
added that has only two purposes: register the devices and map the entire
register addresses. Everything else is left to the dedicated drivers that
will bind to the registered devices.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>