drm/imx: dpu: crtc: Precisely send vbland event if CRTC is active
commit 5316cb3c7b022588520b6b9fb6bc5fa08df0df27 from
https://github.com/nxp-imx/linux-imx

If CRTC is active, we should send vblank event in vblank
interrupt handler to make sure it's sent precisely.  This
patch caches the event to be sent at dpu_crtc->event in
the ->atomic_enable() and the ->atomic_flush() callbacks
and finally sends it out in dpu_vbl_irq_handler().  Since
we rely on the interrupt handler to send the event, we
call drm_crtc_vblank_get() to get a vblank refcount to
guarantee the interrupt is enabled when caching the event
in dpu_crtc_queue_state_event() and call drm_crtc_vblank_put()
to drop a vblank refcount in the interrupt handler.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>