staging: media: imx: enable ISI for imx8mn platform
commit 676d51a2b263171410ba6207a9a8d2a07dd3c7f2 from
https://source.codeaurora.org/external/imx/linux-imx

ISI is image sensor interface of imx8 family. It's reused in imx8mn
platform. But they use different clock tree, so driver need to select
the related clock operation for different platform. In order to solve
the problem, driver define mxc_isi_dev_ops which used to control clock
operation.

Dispmix subsystem of imx8mn is consist of ISI, CSI, DSI and LCDIF modules.
It use GPR to manage the reset and clock signal for all modules. We add
a reset driver for dispmix reset function, so add related consumer of reset
in ISI core driver.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>