gpu: imx: dpu: Access regs in display engine units wo holding mutex
commit 56ec444732b880e5b9016f73472e5685d533e0ee from
https://github.com/nxp-imx/linux-imx

We don't need holding mutex when accessing registers in display engine
units, because KMS is the only relevant client driver and it has ww mutex
mechansim to ensure there is no race condition on the CRTC resources.
Also, we are naturally safe when the driver initializes the units at the
probe and system power management stages.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>