drm/imx: dpu: kms: Support full screen CRTC background
commit 2a2e9bc400292c094d3f40f5a6905b9969d85ee8 from
https://github.com/nxp-imx/linux-imx.git

The CRTC background should be full screen instead of partial
screen, because the DRM core is likely to add configurable
background color support in the future.  We may cover the full
screen with ConstFrame0/1, upon which builds planes.  With this,
it is easier to compute each plane's layer offset vs CRTC start
point and all ConstFrame units can be controlled by CRTC.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>