LF-39 soc: imx: Update busfreq to support different frequncy setpoint
commit 95409ab37680a9f4299a4d9f267e1f501b0e565b from
https://source.codeaurora.org/external/imx/linux-imx

On i.MX8M SOC family, we can support LPDDR4, DDR4 or DDR3L, we may need
to support different setpoint for audio & low bus mode on different DDR type,
So update the code to get all the supported setpoint info from ATF.
The maximum setpoints that can be supported by hardware is 4, if the drate
for a setpoint is '0', that means this setpoint is not enabled. We can use
these info to find out the lowest drate setpoint for audio & low bus mode.

    BuildInfo:
        - ATF 59fe78cfe7

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Anson Huang <Anson.Huang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>