drm/bridge: adv7511: Fix of_reconfig
commit dcdc4c3fa8876156384c6f47911abdf78ca7f488 from
https://github.com/nxp-imx/linux-imx.git

Currently, if the actual i2c device chip (ADV75**) is not found, the
remote node is disabled completely. This causes undefined behavior, if
that device was probed and being used.
In order to fix this behavior, don't disable the remote device, just
detach the remote's endpoint, so that the remote device will just
abandon the connection.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>