LF-824: fw: imx: seco_mu: Add driver for SECO
commit cbe5e27a90341a3c43c9457554c47475d0712d7b from
https://source.codeaurora.org/external/imx/linux-imx

This patch adds the driver to interact with the different
APIs exposed by the SECO using a shared Messaging Unit.

The driver exposes some char devices in user space allowing
a user to send message to the SECO and read its response.

The driver uses the mailbox framework instead of directly
configuring the MU.

Signed-off-by: Silvano di Ninno <silvano.dininno@nxp.com>
Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Acked-by: Dong Aisheng <aisheng.dong@nxp.com>
Acked-by: Leonard Crestez <leonard.crestez@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>