MLK-17077 usb: typec: clear vbus change event in irq handler
commit 99a6592e3a6e19d8b7af558843e37d210a7109ca from
https://github.com/nxp-imx/linux-imx.git

For vbus change event, we need read the vbus status to clear
the alert. Current code do this in queue work, this has problem
on single core running, the queue work of vbus change may have
no chance to be scheduled as we continue receive the vbus change
event in threaded irq.

Acked-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>