mtd: rawnand: Add jedec_id to struct onfi_params
commit 8184f2c98efa18401a06cb2d484cdba60f5069cf from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

During the write oob operation the core will reset the NAND flash,
which switches the flash device to SDR mode from its previously
configured mode but leaves controller in its previous configured
mode which causes wrong clock synchronization between controller
and flash device leads to data mismatch.
The core is issuing the reset because of hardware bugs in some
of the chips like TOSHIBA.
So limit this reset only to TOSHIBA parts. inorder to do that
we need jedec_id. so add jedec_id to struct onfi_params

We don't even need this update if we have nand_setup_data_interface()
hook. but we don't support this because this interface is not supporting
high speed modes.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>