arch: arm64: Fix reserved memory area property name for DSP
commit 7cbcf438df327916f4fd20a8b02dbd581054f48e from
https://github.com/nxp-imx/linux-imx

After change  e00f0e9f115b604 ("ASoC: fsl_dsp: Rename reserved-region
with memory-region") reserved memory area for DSP is now grabbed
from DSP node property named memory-region instead of reserved-region.

We've done this change in order to align with upstream implementation.
Usually we should support the old property name but because
imx-audio-framework is is obsolete and planned for removal in the future
we won't do that.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>