clocksource: timer-keystone: Add platform probe support to existing code
commit c5808acef9de9a4b76fc936f69ff8b57202eac02 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This patch fixes the case for k2g where in clocks are not
available at the time of keystone_timer_init function call.
K2HK/K2L/K2E devices are left untouched and the old way of
keystone_timer_init is preserved since it is init order sensitive.

On top of the existing code a platform driver is introduced so that
if clock is not yet ready probe can be deferred. This enables hres
timers for k2g.

Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>