LF-425-2 drm/bridge: sec-dsim: use 'attach-bridge' for bridge attach
commit b3cf5a11495ed9f1ed36f11cf043b27e8e6d9ebd from
https://source.codeaurora.org/external/imx/linux-imx

Since the port nodes under dsim device can be detached
dynamically by other devices. So use the 'attach-bridge'
property to imply that if there is any available bridge
endpoint node exists. If not, the panel node existence
is required to be detected. If still not, whole display
components binding should be ended which can avoid the
endless defer probe problem when no available display
devices can be attached to dsim.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>