dmaengine: xilinx: Add vdmatest test client code
commit ab760699820df369f94bcbbec0ce4b4e53fa2ae9 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The code is used for axivdma validation. It is using dma_slave_config
generic API and of_dma framework in dma test client with support to
read number of frame stores from device node.

Driver is capable to test multiple instances if present
in the hardware design.

Make the horizontal and vertical length configurable. This allows
VDMA IP to be validated for all supported resolution.

Avoid enabling circular_park mode for S2MM as it might happen that engine
continuously circles through frame buffers w/o being programmed and lead
to an undesired hang. In loopback it is recommended to use same park mode
configuration for both channels.

Signed-off-by: Appana Durga Kedareswara Rao <appana.durga.rao@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radheys@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>