soc: fsl: dpio: fix qbman alignment error in the virtualization context
commit 0e41a31c508493a494084bad732abb5cb5510639 from
https://source.codeaurora.org/external/imx/linux-imx

When running as a guest, under KVM, the CENA region is mapped
as device memory, so uncacheable. When the memory is mapped
as device memory, the unaligned accesses are not allowed.
Memcpy is optimized to transfer 8 bytes at a time regardless
of the start address and might cause alignment issues.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>