soc: fsl: dpio: fix qbman alignment error in the virtualization context
commit c2e094d47712746d75af495f9560fc65b41d0290 from
https://github.com/nxp-imx/linux-imx.git

When running as a guest, under KVM, the CENA region is mapped
as device memory, so uncacheable. When the memory is mapped
as device memory, the unaligned accesses are not allowed.
Memcpy is optimized to transfer 8 bytes at a time regardless
of the start address and might cause alignment issues.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>