LF-221: staging: media: imx: fix potential memory out of bound
commit 0a868886172f777d69945d3175bef07c0cb29254 from
https://github.com/nxp-imx/linux-imx

ISI will not do scale when output width or height equal to input
width or height. But actully, in this case, the source image size
is larger than target size. For instance, 640x480 resize to 640x400.
ISI need to skip scaling operation only when output width and height
equeal to input width and height, so change the OR logic into AND logic.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Sandor.yu <sandor.yu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>