LF-221: staging: media: imx: fix potential memory out of bound
commit d103918b201896574a41a2fe35271db7004854e2 from
https://github.com/nxp-imx/linux-imx.git

ISI will not do scale when output width or height equal to input
width or height. But actully, in this case, the source image size
is larger than target size. For instance, 640x480 resize to 640x400.
ISI need to skip scaling operation only when output width and height
equeal to input width and height, so change the OR logic into AND logic.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Sandor.yu <sandor.yu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>