sdk_dpaa: ceetm: fix recursive dependencies
commit 9c92ed0c46bce3e8b24ecc0487ec5eb975cf91a8 from
https://source.codeaurora.org/external/imx/linux-imx

Due to dependencies between the fsl_ceetm and fsl_dpa modules, remove
the module support for the ceetm driver and integrate it into the main
DPAA Ethernet driver.

The registration of the CEETM Qdisc was the only operation done at
module init. Pass the management of the Qdisc register and unregister
operations to the loading and unloading of the DPAA driver.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>