staging: vchiq_arm: Give vchiq children DT nodes
commit  2f76faa907004e396a68c0641be2b60b0253e891 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

vchiq kernel clients are now instantiated as platform drivers rather
than using DT, but the children of the vchiq interface may still
benefit from access to DT properties. Give them the option of a
a sub-node of the vchiq parent for configuration and to allow
them to be disabled.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>